+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import "../style/global.css";
 import ScrollBtn from "./(common)/scrollBtn";
-import Nav from "./(home)/nav";
+import Nav from "./(common)/nav";
 import ReduxProvider from "../provider/redux";
 import QueryProvider from "@/provider/query";
 import SessionProvider from "@/provider/session";
